Move state and flux to exchange api
